*: support running unit test on Windows
