+++ pkg/utils/storage_unix.go
@@ -1,0 +1,55 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build !windows
+
+package utils
+
+import (
+	"reflect"
+
+	"golang.org/x/sys/unix"
+
+	"github.com/pingcap/dm/pkg/terror"
+)
+
+// GetStorageSize gets storage's capacity and available size
+func GetStorageSize(dir string) (size StorageSize, err error) {
+	var stat unix.Statfs_t
+
+	err = unix.Statfs(dir, &stat)
+	if err != nil {
+		return size, terror.ErrStatFileSize.Delegate(err)
+	}
+
+	// When container is run in MacOS, `bsize` obtained by `statfs` syscall is not the fundamental block size,
+	// but the `iosize` (optimal transfer block size) instead, it's usually 1024 times larger than the `bsize`.
+	// for example `4096 * 1024`. To get the correct block size, we should use `frsize`. But `frsize` isn't
+	// guaranteed to be supported everywhere, so we need to check whether it's supported before use it.
+	// For more details, please refer to: https://github.com/docker/for-mac/issues/2136
+	bSize := uint64(stat.Bsize)
+	field := reflect.ValueOf(&stat).Elem().FieldByName("Frsize")
+	if field.IsValid() {
+		if field.Kind() == reflect.Uint64 {
+			bSize = field.Uint()
+		} else {
+			bSize = uint64(field.Int())
+		}
+	}
+
+	// Available blocks * size per block = available space in bytes
+	size.Available = stat.Bavail * bSize
+	size.Capacity = stat.Blocks * bSize
+
+	return
+}
